% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdiag.R
\name{bdiag}
\alias{bdiag}
\title{bdiag
A function to create block diagonal matrix from list of matrices}
\usage{
bdiag(x)
}
\arguments{
\item{x}{list of adjacency matrices}
}
\description{
bdiag
A function to create block diagonal matrix from list of matrices
}
