% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conn_gnp.R
\name{conn_gnp}
\alias{conn_gnp}
\title{conn_gnp}
\usage{
conn_gnp(n, p, max.it = 10)
}
\arguments{
\item{n}{numebr of nodes per layer}

\item{p}{edge probability}

\item{max.it}{maximum number of iterations to try to get a connected network}
}
\description{
A function to simulate multilayered slowly changing Erdos-Renyi network
}
