% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_norm_laplacian_graph.R
\name{create_norm_laplacian_graph}
\alias{create_norm_laplacian_graph}
\title{create_norm_laplacian_graph}
\usage{
create_norm_laplacian_graph(A)
}
\arguments{
\item{A}{adjacency/supra adjacency matrix}
}
\description{
A function to obtain normalized laplacian
}
