% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_conn_ER_ML.R
\name{get_conn_ER_ML}
\alias{get_conn_ER_ML}
\title{get_ER_ML}
\usage{
get_conn_ER_ML(p, n, tt, corr = 0.98)
}
\arguments{
\item{p}{edge probability}

\item{n}{numebr of nodes per layer}

\item{tt}{number of time layers}

\item{corr}{pearson correlation between layers}
}
\description{
A function to simulate multilayered slowly changing Erdos-Renyi network
}
