% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_supra_adj.R
\name{get_supra_adj}
\alias{get_supra_adj}
\title{get supra adjacency}
\usage{
get_supra_adj(G, periodic = TRUE, omega = 1)
}
\arguments{
\item{G}{list of graphs}

\item{periodic}{whether to have periodic supra adj}

\item{omega}{inter-layer weight}
}
\description{
A function to get supra adjacency from list of adjacency matrices
}
