% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_conn_gnp.R
\name{sample_conn_gnp}
\alias{sample_conn_gnp}
\title{conn_gnp}
\usage{
sample_conn_gnp(G, corr, p, max.it = 10)
}
\arguments{
\item{G}{original graph}

\item{corr}{Pearson correlation between old and new graph}

\item{p}{probability of edge from old network}

\item{max.it}{maximum number of iterations to try to get a connected network}
}
\description{
A function to simulate Erdos-Renyi graph from another graph
}
